requirements.txt
requests>=2.31.0
PyYAML>=6.0

langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0

chromadb>=0.4.0
tiktoken>=0.5.0
networkx>=3.0.0
tenacity>=8.2.0

elasticsearch==7.9.2

# Additional dependencies discovered from codebase analysis
tqdm>=4.65.0
numpy>=1.21.0
scikit-learn>=1.0.0
urllib3>=1.26.0